+++ src/main/java/com/sda/pogodynka/backend/LocationController.java
@@ -8,7 +8,7 @@ public class LocationController {
         this.locationService = locationService;
     }
 
-    public String createNewLocation(String city, int height, int width, String region, String country) {
+    public String createNewLocation(String city, float height, float width, String region, String country) {
 
         Location newLocation = locationService.createLocation(city, height, width, region, country);
         return newLocation.toString();
